backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.3.5
PyPDF2==3.0.1
litellm==1.0.3
supabase==2.0.0
celery==5.3.4
redis==5.0.1
httpx==0.25.2
